+++ packages/gatsby/src/schema/infer/index.js
@@ -1,6 +1,7 @@
 const report = require(`gatsby-cli/lib/reporter`)
 const { ObjectTypeComposer } = require(`graphql-compose`)
-const { getExampleObject, hasNodes } = require(`./inference-metadata`)
+const { hasNodes } = require(`./inference-metadata`)
+const { getExampleObject } = require(`./build-example-data`)
 const { addNodeInterface } = require(`../types/node-interface`)
 const { addInferredFields } = require(`./add-inferred-fields`)
 
@@ -14,8 +15,9 @@ const addInferredTypes = ({
 }) => {
   // XXX(freiksenet): Won't be needed after plugins set typedefs
   // Infer File first so all the links to it would work
-  const typesWithNodes = Object.keys(inferenceMetadata).filter(typeName =>
-    hasNodes(inferenceMetadata[typeName])
+  const { typeMap } = inferenceMetadata
+  const typesWithNodes = Object.keys(typeMap).filter(typeName =>
+    hasNodes(typeMap[typeName])
   )
   const typeNames = putFileFirst(typesWithNodes)
   const noNodeInterfaceTypes = []
@@ -89,14 +91,14 @@ const addInferredType = ({
   // node type owner information directly from store
   if (
     typeComposer.getExtension(`createdFrom`) === `inference` &&
-    hasNodes(inferenceMetadata[typeName])
+    hasNodes(inferenceMetadata.typeMap[typeName])
   ) {
     const nodes = nodeStore.getNodesByType(typeName)
     typeComposer.setExtension(`plugin`, nodes[0].internal.owner)
   }
 
   const exampleValue = getExampleObject({
-    ...inferenceMetadata[typeName],
+    ...inferenceMetadata.typeMap[typeName],
     typeName,
     typeConflictReporter,
   })